+++ package.json
@@ -75,6 +75,7 @@
     "lint-staged": "^11.0.0",
     "postcss": "^8.4.24",
     "prettier": "^2.3.0",
+    "tailwind-scrollbar": "^3.0.4",
     "tailwindcss": "^3.3.2",
     "typescript": "^5.0.4"
   }
