+++ src/components/App/App.tsx
@@ -5,7 +5,7 @@ import nextCookies from 'next-cookies';
 import Router, { useRouter } from 'next/router';
 import FeaturePanel from '../FeaturePanel/FeaturePanel';
 import Map from '../Map/Map';
-import SearchBox from '../SearchBoxNew/SearchBox';
+import SearchBox from '../SearchBox/SearchBox';
 import { MapStateProvider, useMapStateContext } from '../utils/MapStateContext';
 import { getInitialMapView, getInititalFeature } from './helpers';
 import { HomepagePanel } from '../HomepagePanel/HomepagePanel';
