+++ src/components/SearchBox/SearchBox.tsx
@@ -1,107 +1,127 @@
-import React, { useRef, useState } from 'react';
-import styled from 'styled-components';
-import throttle from 'lodash/throttle';
-import SearchIcon from '@material-ui/icons/Search';
-import Paper from '@material-ui/core/Paper';
-import IconButton from '@material-ui/core/IconButton';
+import React, { useState, useEffect } from 'react';
 import Router from 'next/router';
-import { fetchJson } from '../../services/fetch';
+// headlessUI
+import { Combobox } from '@headlessui/react';
+// icons
+import { XMarkIcon, MagnifyingGlassIcon } from '@heroicons/react/24/solid';
+// context
 import { useMapStateContext } from '../utils/MapStateContext';
 import { useFeatureContext } from '../utils/FeatureContext';
-import { AutocompleteInput } from './AutocompleteInput';
+// helpers
 import { t } from '../../services/intl';
-import { ClosePanelButton } from '../utils/ClosePanelButton';
-import { isDesktop, useMobileMode } from '../helpers';
+import { useMobileMode } from '../helpers';
+import { fetchChoices } from './searchBoxUtils';
+import { onHighlightFactory, onSelectedFactory } from './onSelectedFactory';
+// components
+import SearchChoice from './SearchChoice';
 
-const TopPanel = styled.div`
-  position: absolute;
-  height: 72px;
-  box-shadow: 0 10px 20px 0 rgba(0, 0, 0, 0.12);
-  background-color: ${({ theme }) => theme.palette.background.searchBox};
-  padding: 10px;
-  box-sizing: border-box;
-
-  z-index: 1200;
-
-  width: 100%;
-  @media ${isDesktop} {
-    width: 410px;
-  }
-`;
-
-const StyledPaper = styled(Paper)`
-  padding: 2px 4px;
-  display: flex;
-  align-items: center;
-
-  .MuiAutocomplete-root {
-    flex: 1;
-  }
-`;
-
-const SearchIconButton = styled(IconButton)`
-  svg {
-    transform: scaleX(-1);
-    filter: FlipH;
-    -ms-filter: 'FlipH';
-  }
-`;
-
-const getApiUrl = (inputValue, view) => {
-  const [zoom, lat, lon] = view;
-  const lvl = Math.max(0, Math.min(16, Math.round(zoom)));
-  const q = encodeURIComponent(inputValue);
-  return `https://photon.komoot.io/api/?q=${q}&lon=${lon}&lat=${lat}&zoom=${lvl}`;
-};
-
-const fetchOptions = throttle(async (inputValue, view, setOptions) => {
-  const searchResponse = await fetchJson(getApiUrl(inputValue, view));
-  const options = searchResponse.features;
-  setOptions(options || []);
-}, 400);
+interface Choice {
+  id: string;
+  properties: {
+    name: string;
+  };
+}
 
 const SearchBox = () => {
-  const { featureShown, feature, setFeature, setPreview } = useFeatureContext();
   const { view } = useMapStateContext();
-  const [inputValue, setInputValue] = useState('');
-  const [options, setOptions] = useState([]);
-  const autocompleteRef = useRef();
   const mobileMode = useMobileMode();
+  const { featureShown, feature, setFeature, setPreview } = useFeatureContext();
 
-  React.useEffect(() => {
-    if (inputValue === '') {
-      setOptions([]);
-      return;
-    }
-    fetchOptions(inputValue, view, setOptions);
-  }, [inputValue]);
+  const [query, setQuery] = useState('');
+  const [choices, setChoices] = useState<Choice[]>([]);
+  const [selectedChoice, setSelectedChoice] = useState<Choice | null>(null);
 
-  const closePanel = () => {
-    setInputValue('');
+  function onInputChange(event) {
+    setQuery(event.target.value);
+    onHighlightFactory(setPreview, {});
+  }
+
+  // close feature panel button (clear selection basically)
+  function closePanel() {
+    setQuery('');
     if (mobileMode) {
       setPreview(feature);
     }
     setFeature(null);
     Router.push(`/${window.location.hash}`);
-  };
+  }
+
+  // on query change
+  useEffect(() => {
+    // reset if query is empty
+    if (query === '') {
+      setChoices([]);
+      setSelectedChoice(null);
+      return;
+    }
+    // fetch choices
+    fetchChoices(query, view, setChoices);
+  }, [query]);
+
+  // when selected changes
+  useEffect(() => {
+    if (selectedChoice === null) return;
+    onSelectedFactory(setFeature, setPreview, mobileMode, selectedChoice);
+  }, [selectedChoice]);
 
   return (
-    <TopPanel>
-      <StyledPaper elevation={1} ref={autocompleteRef}>
-        <SearchIconButton disabled aria-label={t('searchbox.placeholder')}>
-          <SearchIcon />
-        </SearchIconButton>
+    <Combobox value={selectedChoice} onChange={setSelectedChoice}>
+      {/* input container */}
+      <div className="relative">
+        {/* search icon */}
+        {/* magnifying glass icon */}
+        <div className="absolute top-0 bottom-0 left-0 flex items-center justify-center pointer-events-none text-zinc-400 aspect-square">
+          <MagnifyingGlassIcon className="h-6 w-6" />
+        </div>
 
-        <AutocompleteInput
-          inputValue={inputValue}
-          setInputValue={setInputValue}
-          options={options}
-          autocompleteRef={autocompleteRef}
+        {/* input */}
+        <Combobox.Input
+          className="bg-white dark:bg-zinc-800 shadow-md shadow-black/20 rounded-lg placeholder-zinc-500 dark:placeholder-zinc-400 text-zinc-800 dark:text-zinc-200 pl-12 w-full h-12 outline-0 transition truncate"
+          // @ts-ignore
+          placeholder={t('searchbox.placeholder')}
+          displayValue={(choice: Choice | null) => {
+            if (choice === null) return query;
+            return choice.properties.name;
+          }}
+          onChange={onInputChange}
+          spellCheck="false"
         />
 
-        {featureShown && <ClosePanelButton onClick={closePanel} />}
-      </StyledPaper>
-    </TopPanel>
+        {/* close button */}
+        <div className="absolute top-0 bottom-0 right-0 aspect-square">
+          {featureShown && (
+            <button
+              type="button"
+              onClick={closePanel}
+              className="flex items-center justify-center w-full h-full"
+            >
+              <span className="sr-only">Close</span>
+              <XMarkIcon className="h-6 w-6 text-zinc-700 dark:text-zinc-400" />
+            </button>
+          )}
+        </div>
+      </div>
+
+      {/* Choices */}
+      <Combobox.Options className="rounded-md overflow-y-auto shadow-md shadow-black/20 z-10 scrollbar-thin scrollbar-track-white dark:scrollbar-track-zinc-800 scrollbar-thumb-zinc-300 dark:scrollbar-thumb-zinc-600 scrollbar-thumb-rounded-full">
+        {choices.map((choice) => (
+          <Combobox.Option key={choice.id} value={choice}>
+            {/* TODO replace with a new component */}
+            {({ active }) => (
+              <li
+                className={`flex justify-between p-2 transition duration-10 ${
+                  active
+                    ? 'bg-zinc-200 dark:bg-zinc-700'
+                    : 'bg-white dark:bg-zinc-800'
+                }`}
+              >
+                <SearchChoice query={query} choice={choice} />
+              </li>
+            )}
+          </Combobox.Option>
+        ))}
+      </Combobox.Options>
+    </Combobox>
   );
 };
