+++ src/components/SearchBox/SearchChoice.jsx
