+++ src/components/SearchBox/searchBoxUtils.js
